requirements.txt
  lyricsgenius==3.6.2
  openai==1.77.0
  pydbus==0.6.0
  python-dotenv==1.1.0
  requests==2.32.3
  rich==14.0.0
  spotipy==2.25.1
tiktoken==0.9.0
pylast==5.2.0
